% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_demo.R
\name{kmeRs_demo}
\alias{kmeRs_demo}
\title{Simple Demo For KmeRs Package}
\usage{
kmeRs_demo()
}
\value{
example report
}
\description{
The \code{kmeRs_demo} function calculates and shows an example report of
the kmeRs package for sample given k-mers: "ATA", "CGC", "TGC", "GGA"
}
\examples{
# Test package - example raport

kmeRs_demo()

}
