% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_similarity_matrix.R
\name{kmeRs_generate_kmers}
\alias{kmeRs_generate_kmers}
\title{kmeRs generate kmers}
\usage{
kmeRs_generate_kmers(k, bases)
}
\arguments{
\item{k}{times}

\item{bases}{follow the kmeRs_similarity_matrix()}
}
\description{
kmeRs generate kmers
}
