% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_score_and_sort.R
\name{kmeRs_score_and_sort}
\alias{kmeRs_score_and_sort}
\title{Score And Sort The Similarity Matrix}
\usage{
kmeRs_score_and_sort(kmeRs_similarity_matrix)
}
\arguments{
\item{kmeRs_similarity_matrix}{the similarity matrix calculated by \code{kmeRs_similarity_matrix} function}
}
\value{
sorted similarity matrix with global.score column added; is returned as a data.frame
}
\description{
The \code{kmeRs_score_and_sort} function sums the partial scores and sort the data.frame
to indicate the most 'different' k-mers
}
\examples{
# Calculate the example BLOSUM62 matrix and score the result

example <- kmeRs_similarity_matrix(kmers_given = c("A", "T", "C", "G"), submat = "BLOSUM62")
kmeRs_score_and_sort(example)

}
