% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_show_alignment.R
\name{kmeRs_show_alignment}
\alias{kmeRs_show_alignment}
\title{Calculate and Show Alignment Between Two Compared K-mers}
\usage{
kmeRs_show_alignment(kmer_A = "", kmer_B = "", submat = "BLOSUM62")
}
\arguments{
\item{kmer_A}{given k-mer A}

\item{kmer_B}{given k-mer B}

\item{submat}{substitution matrix version, default is 'BLOSUM62'; could be one of 'BLOSUM45', 'BLOSUM50', 'BLOSUM62', 'BLOSUM80',
'BLOSUM100', 'PAM30', 'PAM40', 'PAM70', 'PAM120', 'PAM250'}
}
\value{
alignment is returned as a data.frame
}
\description{
The \code{kmeRs_show_alignment} function aligns and shows calculated alignment between two DNA or RNA sequences
}
\examples{
# Example alignment

kmeRs_show_alignment( kmer_A = "AAATTTCCCGGG", kmer_B = "TCACCC" ,submat = "BLOSUM62")

}
