% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeRs_transcript_to_complementary.R
\name{kmeRs_transcript_to_complementary}
\alias{kmeRs_transcript_to_complementary}
\title{Translate Given K-mers To Complementary Sequences}
\usage{
kmeRs_transcript_to_complementary(kmers_given)
}
\arguments{
\item{kmers_given}{vector contains given k-mers}
}
\value{
vector contains complementary sequences
}
\description{
The \code{kmeRs_transcript_to_complementary} function transcripts DNA given k-mers to complementary sequences
}
\examples{
# Returns complementary sequence to GATTACA

kmeRs_transcript_to_complementary('GATTACA')

}
